+++ users/src/main/kotlin/org/seasonist/dtos/ErrorResponse.kt
@@ -1,4 +1,4 @@
-package org.seasonist.dtos.request
+package org.seasonist.dtos
 
 data class ErrorResponse(
 	val message: String
